+++ xstampp.astpa/src/xstampp/astpa/model/controlaction/ControlActionController.java
@@ -17,7 +17,6 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Observable;
 import java.util.UUID;
 
 import javax.xml.bind.annotation.XmlAccessType;
@@ -30,6 +29,7 @@ import org.eclipse.core.runtime.Assert;
 
 import messages.Messages;
 import xstampp.astpa.model.ATableModel;
+import xstampp.astpa.model.ATableModelController;
 import xstampp.astpa.model.controlaction.interfaces.IControlAction;
 import xstampp.astpa.model.controlaction.interfaces.IUnsafeControlAction;
 import xstampp.astpa.model.controlaction.interfaces.UnsafeControlActionType;
@@ -43,8 +43,6 @@ import xstampp.astpa.model.linking.Link;
 import xstampp.astpa.model.linking.LinkController;
 import xstampp.astpa.model.linking.LinkingType;
 import xstampp.astpa.model.service.UndoCSCChangeCallback;
-import xstampp.astpa.model.service.UndoControlActionChangeCallback;
-import xstampp.astpa.model.service.UndoUCAChangesCallback;
 import xstampp.model.AbstractLTLProvider;
 import xstampp.model.IEntryFilter;
 import xstampp.model.NumberedArrayList;
@@ -58,7 +56,7 @@ import xstampp.ui.common.ProjectManager;
  * 
  */
 @XmlAccessorType(XmlAccessType.NONE)
-public class ControlActionController extends Observable implements IControlActionController {
+public class ControlActionController extends ATableModelController implements IControlActionController {
 
   @XmlElementWrapper(name = "controlactions")
   @XmlElement(name = "controlaction")
@@ -269,24 +267,8 @@ public class ControlActionController extends Observable implements IControlActio
 
   @Override
   public boolean setUcaDescription(UUID unsafeControlActionId, String description) {
-    if ((unsafeControlActionId == null) || (description == null)) {
-      return false;
-    }
-    String oldDescription = null;
-    UnsafeControlAction unsafeControlAction = (UnsafeControlAction) this
-        .getUnsafeControlAction(unsafeControlActionId);
-    if (unsafeControlAction != null) {
-      oldDescription = unsafeControlAction.setDescription(description);
-    }
-    if (oldDescription != null) {
-      setChanged();
-      UndoUCAChangesCallback callback = new UndoUCAChangesCallback(this, unsafeControlActionId);
-      callback.setDescriptionChange(oldDescription, description);
-
-      notifyObservers(callback);
-      return true;
-    }
-    return false;
+    return setModelDescription(getUnsafeControlAction(unsafeControlActionId), description,
+        ObserverValue.UNSAFE_CONTROL_ACTION);
   }
 
   @Override
@@ -354,46 +336,12 @@ public class ControlActionController extends Observable implements IControlActio
 
   @Override
   public boolean setControlActionDescription(UUID controlActionId, String description) {
-    if ((controlActionId == null) || (description == null)) {
-      return false;
-    }
-    ITableModel controlAction = getControlAction(controlActionId);
-    if (!(controlAction instanceof ControlAction)) {
-      return false;
-    }
-    String oldDescription = ((ControlAction) controlAction).setDescription(description);
-    if (oldDescription != null) {
-      setChanged();
-      UndoControlActionChangeCallback changeCallback = new UndoControlActionChangeCallback(this,
-          controlAction);
-      changeCallback.setDescriptionChange(oldDescription, description);
-      notifyObservers(changeCallback);
-      return true;
-    }
-    return false;
+    return setModelDescription(getControlAction(controlActionId), description, ObserverValue.CONTROL_ACTION);
   }
 
   @Override
   public boolean setControlActionTitle(UUID controlActionId, String title) {
-    if ((controlActionId == null) || (title == null)) {
-      return false;
-    }
-    ITableModel controlAction = getControlAction(controlActionId);
-    if (!(controlAction instanceof ControlAction)) {
-      return false;
-    }
-    boolean result = false;
-
-    String oldTitle = ((ControlAction) getControlAction(controlActionId)).setTitle(title);
-    if (oldTitle != null) {
-      setChanged();
-      UndoControlActionChangeCallback changeCallback = new UndoControlActionChangeCallback(this,
-          controlAction);
-      changeCallback.setTitleChange(oldTitle, title);
-      notifyObservers(changeCallback);
-      return true;
-    }
-    return result;
+    return setModelTitle(getControlAction(controlActionId), title, ObserverValue.CONTROL_ACTION);
   }
 
   @Override